requirements.txt
PyDrive==1.3.1
pytest
python_pidfile==3.0.0
requests==2.21.0
SQLAlchemy==0.9.10
pandas==1.3.3
mutagen==1.45.1
flask_restplus==0.13.0
httplib2==0.15.0
importlib_metadata==3.10.1
beautifulsoup4==4.10.0
Flask==2.0.2
flask-cors==3.0.10
flask-socketio==4.3.2
python-socketio==4.6.1
python-engineio==3.14.2
flask-httpauth
google_api_python_client==2.33.0
jsonpickle==2.0.0
pidfile==0.1.1
python_dateutil==2.8.2
tinycss2==1.1.1
selenium
chromedriver_autoinstaller
python-dotenv
pyyaml
keyring
fuzzywuzzy
flask-caching==1.10.1
ffmpeg-python==0.2.0
portalocker
pillow
pytest-mock
requests-mock